Split big files into smaller ones
